refactor(checker): extract shared temp bundler runner
Extract duplicated temp-directory lifecycle and bundler execution plumbing into TempBundleRunner.

Both checkers now delegate the common flow (tmp dir creation, temporary Gemfile setup, bundler definition/install, and stdout capture) while preserving checker-specific gemfile content and installer options.

This is a no-behavior-change refactor intended to reduce duplication and make future checker changes less error-prone.
